+++ Source/Patches/CarPatches.cs
@@ -1,5 +1,10 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection.Emit;
 using HarmonyLib;
 using LCVR.Player;
+using Unity.Netcode;
+using UnityEngine;
 
 namespace LCVR.Patches;
 
@@ -37,5 +42,46 @@ internal static class VRCarPatches
             wheel.GetSteeringInput();
         
         return false;
+    }
+
+    /// <summary>
+    /// Remove player body animations from car update loop as we are controlling these values ourselves
+    /// </summary>
+    [HarmonyPatch(typeof(VehicleController), nameof(VehicleController.Update))]
+    [HarmonyTranspiler]
+    private static IEnumerable<CodeInstruction> RemovePlayerAnimations(IEnumerable<CodeInstruction> instructions)
+    {
+        return new CodeMatcher(instructions)
+            .MatchForward(false, [new CodeMatch(OpCodes.Ldstr, "SA_CarAnim")])
+            .Advance(-23)
+            // Keep single nop since this instruction can be jumped to
+            .SetOpcodeAndAdvance(OpCodes.Nop)
+            .RemoveInstructions(39)
+            .InstructionEnumeration();
+    }
+
+    /// <summary>
+    /// If we are the host, and we're dead, move the ownership to the first alive player to not interfere with the collision
+    /// </summary>
+    [HarmonyPatch(typeof(VehicleController), nameof(VehicleController.RemovePlayerControlOfVehicleServerRpc))]
+    [HarmonyPostfix]
+    private static void OnRemoveDriver(VehicleController __instance)
+    {
+        if (!NetworkManager.Singleton.IsHost)
+            return;
+
+        if (!VRSession.Instance.LocalPlayer.PlayerController.isPlayerDead)
+            return;
+
+        var alivePlayer = StartOfRound.Instance.allPlayerScripts.FirstOrDefault(player => !player.isPlayerDead);
+        if (!alivePlayer)
+            return;
+
+        __instance.syncCarPositionInterval = 1f;
+        __instance.syncedPosition = Vector3.zero;
+        __instance.syncedRotation = Quaternion.identity;
+        __instance.SyncCarPhysicsToOtherClients();
+
+        __instance.NetworkObject.ChangeOwnership(alivePlayer.actualClientId);
     }
 }