+++ CHANGELOG.md
@@ -66,6 +66,7 @@ Some bindings are blacklisted from being used, mostly the "touched" bindings on 
 - Reworked the input system to allow for manual control binding overriding
 - XR Origin now has the same parent as the local player
 - Reworked a small portion of the 6DOF system to allow 6DOF when parented to other objects
+- Resetting height now also resets your rotation if you are in a special interact animation (i.e. terminal, car, enemy kill animation)
 
 **Removals**: